Fix the five port defects, and correct the docs they invalidated
All five tests from the previous commit now pass.

  1. Integer.IntValue narrows through int32. Java (int) truncates; Go
     int(int64) does not, because int is 64 bits on every platform this
     builds for. This also repairs JAVA-BUGS entry 1: Equals was supposed
     to reproduce the 32-bit truncation and, because IntValue did not
     narrow, it was not doing so.
  2. Float.IntValue and LongValue saturate. Java narrowing from a floating
     point type clamps to MAX_VALUE or MIN_VALUE and maps NaN to zero; Go
     leaves an out-of-range conversion undefined and was yielding MinInt64
     for both signs of MaxFloat32.
  3. Flate.Decode logs corrupt input instead of propagating it, and hands
     back what inflated. This is the behaviour the filter exists for: the
     Java comment reads "do not throw an exception, use the already read
     data or an empty stream". The port had preserved the checksum-bypass
     half of that damage tolerance and dropped this half.
  4. Stream.Length returns ErrStreamWriting while a writer is open, where
     Java throws IllegalStateException. The length is only recorded on
     close, so the previous code returned a stale value silently. Signature
     is now (int64, error).
  5. XrefTrailerResolver defaults a resolved section to XRefTypeTable,
     matching the Java constructor, so a file whose startxref points
     nowhere still reports TABLE rather than NONE.

Also corrected: the xref Compare doc claimed a nil entry sorts last when
it sorts first, and STATUS.md still listed Integer.Equals, ParseHexString
and assertBytesEqual as deviations after they were reverted to match the
Java. STATUS.md now lists those four under behaviour that looks like a
deviation and is not.

Worth recording: none of these was caught by the ported tests, even though
those were written from the Java. Each turns on a way Go differs from Java
-- integer width, float conversion, error propagation -- rather than on
what PDFBox does, which is the blind spot of porting test-first from the
source alone.

+++ go/pdfbox/cos/integer.go
@@ -72,8 +72,12 @@ func GetInteger(value int64) *Integer {
 // FloatValue returns the value as a float32.
 func (i *Integer) FloatValue() float32 { return float32(i.value) }
 
-// IntValue returns the value truncated to an int, as Java's (int) cast does.
-func (i *Integer) IntValue() int { return int(i.value) }
+// IntValue returns the value narrowed to 32 bits.
+//
+// Port of intValue(), which is `(int) value`. Java's narrowing cast discards
+// everything above bit 31; Go's int(int64) does not, because int is 64 bits on
+// every platform this builds for, so the truncation has to be written out.
+func (i *Integer) IntValue() int { return int(int32(i.value)) }
 
 // LongValue returns the value.
 func (i *Integer) LongValue() int64 { return i.value }
